+++ resources/views/forms/components/translatable-container.blade.php
@@ -1,12 +1,4 @@
 <div>
-    <style>
-        .translate-svg-color {
-            fill: #000;
-        }
-        html.dark .translate-svg-color {
-            fill: #fff;
-        }
-    </style>
     <div
             x-data="{
                 open: false,
@@ -37,12 +29,12 @@
         <div class="flex items-center gap-1.5 cursor-pointer select-none mt-2"
              @click="handleOpenState()"
         >
-            <div x-show="!open" class="translate-svg-color">
-                <svg xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" viewBox="0 0 24 24" xml:space="preserve" width="24px" height="24px"><g id="SVGRepo_bgCarrier" stroke-width="0"></g><g id="SVGRepo_tracerCarrier" stroke-linecap="round" stroke-linejoin="round"></g><g id="SVGRepo_iconCarrier"> <style type="text/css"> .st0{fill:none;} </style> <path d="M9,18l7-6L9,6V18z"></path> <rect class="st0" width="24" height="24"></rect> <rect class="st0" width="24" height="24"></rect> </g></svg>
+            <div x-show="!open">
+                <x-filament::icon icon="heroicon-c-chevron-right" class="h-5 w-5 text-gray-500 dark:text-gray-400"/>
             </div>
 
-            <div x-show="open" class="translate-svg-color">
-                <svg xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" viewBox="0 0 24 24" xml:space="preserve" width="24px" height="24px"><g id="SVGRepo_bgCarrier" stroke-width="0"></g><g id="SVGRepo_tracerCarrier" stroke-linecap="round" stroke-linejoin="round"></g><g id="SVGRepo_iconCarrier"> <style type="text/css"> .st0{fill:none;} </style> <path d="M6.5,8.5l6,7l6-7H6.5z"></path> <rect class="st0" width="24" height="24"></rect> <rect class="st0" width="24" height="24"></rect> </g></svg>
+            <div x-show="open">
+                <x-filament::icon icon="heroicon-c-chevron-down" class="h-5 w-5 text-gray-500 dark:text-gray-400"/>
             </div>
 
             @foreach($getTranslatableLocales() as $locale)